Clean up and rename.
